refactor: Separate model and user widget
